+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 TBD: Version 1.1 (??? 2018)
 --------------------------------------
 New Rules
-   - #247: MissingOverrideAnnotation rule (enhanced) - Checks for methods that override a method in a super class or implement a method in an interface but are not annotated with @Override. (Marcin Erdmann)
-   - #271: Indentation rule (formatting) - Check indentation for class, method and field declarations
+   - #247: **MissingOverrideAnnotation** rule (enhanced) - Checks for methods that override a method in a super class or implement a method in an interface but are not annotated with @Override. (Marcin Erdmann)
+   - #271: **Indentation** rule (formatting) - Check indentation for class, method and field declarations and initial statements.
 
 Updated/Enhanced Rules and Bug Fixes
    - #239: Ignore super calls in UnnecessarySetterRule. (Marcin Erdmann)